Add new Javy JSON builtins
This commit introduces the usage of the the new `Javy.JSON` builtins.

Also bumps the version and removes the dependency on the `javy` npm
package as nothing from that package is consumed in this version
anymore.

Opted to introduce the TypeScript definition in this package instead of
including a .d.ts file in the `javy` npm package given the reason stated
in the previous paragraph.