Introduce Javy.JSON builtins
This commit introduces `Javy.JSON` builtins and bumps the package
version to 1.0.0.

It's important to note that this is a breaking change because it
introduces new JavaScript builtins for faster JSON parsing and
stringifying from stdin and to stdout.

In order to use v1.0.0 of this package, users need the latest shopify
cli, which contains `javy` v3.1.0